Fix Android sound to properly parse WAV files. Fix memory leak for Android sound handling. Improve error handling for Android sound loading.

+++ androidaudio/androidsoundeffect.h
@@ -43,7 +43,7 @@ public slots:
 
 private:
     char* mBuffer;
-    off_t mLength;
+    int mLength;
 
     QString mPath;
